+++ models/all.go
@@ -19,7 +19,7 @@ type User struct {
 			UserID          int
 			AccountID       int
 			Name            string
-			Value           int64
+			Value           float64
 			TransactionType bool
 			Edited          time.Time
 			Done            time.Time
